stackoverflow-trends/assets/raw/requirements.txt
pandas
requests
python-dateutil
